Closes rawket-vbxq: Bridge library (src/bridge.rs)
Closes rawket-5sz1: Filter expression language (src/filter.rs)

Part of epic rawket-7j1i: VirtualLink late peer binding + test bridge

+++ src/tcp.rs
@@ -2486,7 +2486,7 @@ impl TcpSocket {
     pub fn snd_nxt(&self) -> u32 { self.snd_nxt.as_u32() }
     pub fn snd_una(&self) -> u32 { self.snd_una.as_u32() }
     pub fn rcv_nxt(&self) -> u32 { self.rcv_nxt.as_u32() }
-    pub fn bytes_in_flight(&self) -> u32 { (self.snd_nxt - self.snd_una) as u32 }
+    pub fn bytes_in_flight(&self) -> u32 { self.snd_nxt - self.snd_una }
     pub fn send_buf_len(&self) -> usize { self.send_buf.len() }
     pub fn bbr_delivered(&self) -> u64 { self.bbr.delivered }
     pub fn bbr_next_round_delivered(&self) -> u64 { self.bbr.next_round_delivered }
